Add WiFi firmware for Thinkpad X13s
This is not the correct firmware for the device but it does work.

The upstream ath11k driver needs changes to load our actual firmware so
we have a hack in place that loads this board-2.bin and we fake being a
supported board.  WiFi works, but it's not the greatest performance.
Sometimes 5GHz shows up, sometimes it doesn't.  Even when it does, you
will never see > 30% signal strength.